backend/requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
torch
numpy
python-dotenv